test: migrate traverse.spec.ts to karma suite
