+++ packages/integration-karma/test/shadow-dom/Element-properties/x/customElementAsDefaultSlot/customElementAsDefaultSlot.html
@@ -1,0 +1,5 @@
+<template>
+    <slot>
+        <x-child></x-child>
+    </slot>
+</template>
