+++ packages/integration-karma/test/shadow-dom/Element-properties/x/customElementAsDefaultSlot/customElementAsDefaultSlot.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class MyComponent extends LightningElement {}
