+++ packages/integration-karma/test/shadow-dom/Element-properties/x/noSlot/noSlot.html
@@ -1,0 +1,1 @@
+<template></template>
