+++ packages/integration-karma/test/shadow-dom/Element-properties/x/simpleParent/simpleParent.html
@@ -1,0 +1,4 @@
+<template>
+    <div></div>
+    <x-no-slot></x-no-slot>
+</template>
