+++ packages/integration-karma/test/shadow-dom/Element-properties/x/slottedCustomElement/slottedCustomElement.html
@@ -1,0 +1,5 @@
+<template>
+    <x-slot>
+        <x-child></x-child>
+    </x-slot>
+</template>
