+++ packages/integration-karma/test/shadow-dom/Element-properties/x/slottedCustomElement/slottedCustomElement.js
@@ -1,0 +1,3 @@
+import { LightningElement } from 'lwc';
+
+export default class MyComponent extends LightningElement {}
