+++ packages/integration-karma/test/shadow-dom/Element-properties/x/textSlotted/textSlotted.html
@@ -1,0 +1,5 @@
+<template>
+    <x-slot>
+        text
+    </x-slot>
+</template>
